backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-docx==1.1.0
PyMuPDF==1.23.8
chromadb==0.4.18
numpy==1.24.3
requests==2.31.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0 